Adicionando exclusão lógica na presença e controle de alteraçao apenas para a presença ativa

+++ src/Crossfit/Dados/AlunoAula.php
@@ -31,9 +31,9 @@ class AlunoAula
 	{
 		$sql = 'select aula.id_aula as id_aula, aula.horario as horario, count(alunos_aula.id_aluno) as num_presentes, aula.excedente as num_excedentes from aula
 				left join alunos_aula on aula.id_aula = alunos_aula.id_aula 
-				where aula.data = ? and aula.id_organizacao = ?
+				where aula.data = ? and aula.id_organizacao = ? and aula.status = ?
 				group by aula.id_aula;';
-		$resultado = Conexao::get()->fetchAll($sql, array($data, App::getSession()->get('organizacao')));
+		$resultado = Conexao::get()->fetchAll($sql, array($data, App::getSession()->get('organizacao'), 'A'));
 		return $resultado;
 	}
 
@@ -67,35 +67,40 @@ class AlunoAula
 
 	public static function atualizaPresenca($id_aula, $presencaDataset)
 	{
-		$aulaDataset = array(
-			'data' => $presencaDataset->data,
-			'horario' => $presencaDataset->horario,
-			'excedente' => $presencaDataset->excedente,
-			'id_organizacao' => App::getSession()->get('organizacao')
-		);
+		$aulaAtiva = Aula::retornaIdAulaAtiva();
 
-		Conexao::get()->update('aula',$aulaDataset, array('id_aula' => $id_aula));		
-
-		Conexao::get()->delete('alunos_aula', array('id_aula' => $id_aula, 'id_organizacao' => App::getSession()->get('organizacao')));
-
-		$presentes = $presencaDataset->presentes;
-		foreach ($presentes as $presente) {
-			$alunosAulaDataset = array(
-				'id_aluno' => $presente->id_aluno,
-				'id_aula' => $id_aula,
-				'num_senha' => $presente->senha,
+		if($aulaAtiva == $id_aula) {
+			$aulaDataset = array(
+				'data' => $presencaDataset->data,
+				'horario' => $presencaDataset->horario,
+				'excedente' => $presencaDataset->excedente,
 				'id_organizacao' => App::getSession()->get('organizacao')
 			);
 
-			Conexao::get()->insert('alunos_aula', $alunosAulaDataset);
-		}
+			Conexao::get()->update('aula',$aulaDataset, array('id_aula' => $id_aula));		
 
-		return true;
+			Conexao::get()->delete('alunos_aula', array('id_aula' => $id_aula, 'id_organizacao' => App::getSession()->get('organizacao')));
+
+			$presentes = $presencaDataset->presentes;
+			foreach ($presentes as $presente) {
+				$alunosAulaDataset = array(
+					'id_aluno' => $presente->id_aluno,
+					'id_aula' => $id_aula,
+					'num_senha' => $presente->senha,
+					'id_organizacao' => App::getSession()->get('organizacao')
+				);
+
+				Conexao::get()->insert('alunos_aula', $alunosAulaDataset);
+			}	
+
+			return true;
+		}
+		return false;
 	}
 
 	public static function removeAlunosAula($id_aula)
 	{
-		Conexao::get()->delete('aula', array('id_aula' => $id_aula, 'id_organizacao' => App::getSession()->get('organizacao')));
+		Conexao::get()->update('aula', array('status' => 'I'), array('id_aula' => $id_aula, 'id_organizacao' => App::getSession()->get('organizacao')));
 
 		return true;
 	}
